Update original file

+++ Python/bytecodes.c
@@ -5251,8 +5251,7 @@ dummy_func(
                 printf("SIDE EXIT: [UOp ");
                 _PyUOpPrint(&next_uop[-1]);
                 printf(", exit %tu, temp %d, target %d -> %s]\n",
-                    (unsigned long)(exit - current_executor->exits),
-                    exit->temperature.value_and_backoff,
+                    exit - current_executor->exits, exit->temperature.value_and_backoff,
                     (int)(target - _PyFrame_GetBytecode(frame)),
                     _PyOpcode_OpName[target->op.code]);
             }
